add build_no_clean ability

+++ bin/build.sh
@@ -101,6 +101,9 @@ build_fedora_base_image() {
 }
 
 build_fedora_clean() {
+    if [[ ${build_no_clean:-} ]]; then
+        return
+    fi
     # Caches
     build_yum clean all
     ls -d /var/cache/*/* | grep -v /var/cache/ldconfig/ | xargs rm -rf
